% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAbAtlas-package.R
\docType{package}
\name{AAbAtlas-package}
\alias{AAbAtlas}
\alias{AAbAtlas-package}
\title{AAbAtlas: Access the Antibodyome Atlas of Autoantibody Profiling Datasets}
\description{
Client for the antibodyome-atlas, a curated catalog of autoantibody profiling datasets and papers (protein/antigen microarrays and related platforms) drawn from public repositories such as GEO and ArrayExpress. Metadata is fetched live from the atlas's GitHub-hosted JSON manifest, so results always reflect the latest curated data (or a pinned release, for reproducibility). Provides functions to list and filter datasets, resolve citations and related records, and download the underlying data files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahcm088/antibodyome-atlas}
  \item Report bugs at \url{https://github.com/ahcm088/antibodyome-atlas/issues}
}

}
\author{
\strong{Maintainer}: Alexandre Marques \email{marquesufcg@gmail.com}

}
\keyword{internal}
